+++ tests/e2e/e2e_evm_test.go
@@ -3,15 +3,12 @@ package e2e
 import (
 	"bytes"
 	"context"
-	"crypto/ecdsa"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"log"
 	"math/big"
 	"net/http"
 	"strings"
-	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
@@ -19,10 +16,7 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient"
 
-	"cosmossdk.io/math"
-
 	"github.com/cosmos/cosmos-sdk/crypto/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/bech32"
 
 	"github.com/kiichain/kiichain/v2/tests/e2e/mock"
@@ -33,9 +27,9 @@ const (
 )
 
 // testEVMQueries Test EVM queries
-func (s *IntegrationTestSuite) testEVMQueries(jsonRCP string) {
+func (s *IntegrationTestSuite) testEVMQueries(jsonRPC string) {
 	s.Run("eth_blockNumber", func() {
-		res, err := httpEVMPostJSON(jsonRCP, "eth_blockNumber", []interface{}{})
+		res, err := httpEVMPostJSON(jsonRPC, "eth_blockNumber", []interface{}{})
 		s.Require().NoError(err)
 
 		blockNumber, err := parseResultAsHex(res)
@@ -44,7 +38,7 @@ func (s *IntegrationTestSuite) testEVMQueries(jsonRCP string) {
 	})
 
 	s.Run("eth_chainId", func() {
-		res, err := httpEVMPostJSON(jsonRCP, "eth_chainId", []interface{}{})
+		res, err := httpEVMPostJSON(jsonRPC, "eth_chainId", []interface{}{})
 		s.Require().NoError(err)
 
 		chainID, err := parseResultAsHex(res)
@@ -53,7 +47,7 @@ func (s *IntegrationTestSuite) testEVMQueries(jsonRCP string) {
 	})
 
 	s.Run("eth_getBalance on zero address", func() {
-		res, err := httpEVMPostJSON(jsonRCP, "eth_getBalance", []interface{}{
+		res, err := httpEVMPostJSON(jsonRPC, "eth_getBalance", []interface{}{
 			"0x0000000000000000000000000000000000000000", "latest",
 		})
 		s.Require().NoError(err)
@@ -64,7 +58,7 @@ func (s *IntegrationTestSuite) testEVMQueries(jsonRCP string) {
 	})
 
 	s.Run("web3_clientVersion", func() {
-		res, err := httpEVMPostJSON(jsonRCP, "web3_clientVersion", []interface{}{})
+		res, err := httpEVMPostJSON(jsonRPC, "web3_clientVersion", []interface{}{})
 		s.Require().NoError(err)
 
 		_, ok := res["result"].(string)
@@ -73,45 +67,32 @@ func (s *IntegrationTestSuite) testEVMQueries(jsonRCP string) {
 }
 
 // testEVM Tests EVM send and contract usage
-func (s *IntegrationTestSuite) testEVM(jsonRCP string) {
-	var (
-		err           error
-		valIdx        = 0
-		c             = s.chainA
-		chainEndpoint = fmt.Sprintf("http://%s", s.valResources[c.id][valIdx].GetHostPort("1317/tcp"))
-	)
+func (s *IntegrationTestSuite) testEVM(jsonRPC string) {
+	var err error
 
 	// Get a funded EVM account and check balance transactions
-	key, _ := s.setupEVMwithFunds(jsonRCP, chainEndpoint, valIdx)
+	evmAccount := s.chainA.evmAccount
 
 	// Setup client
-	client, err := ethclient.Dial(jsonRCP)
+	client, err := ethclient.Dial(jsonRPC)
 	s.Require().NoError(err)
 
 	// Deploy contract
 	s.Run("create and interact w/ contract", func() {
 		// Prepare auth
-		auth, err := bind.NewKeyedTransactorWithChainID(key, big.NewInt(1010))
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		// Set optional params
-		auth.Value = big.NewInt(0)
-		auth.GasLimit = uint64(3000000) // gas limit
-		auth.GasPrice, _ = client.SuggestGasPrice(context.Background())
+		auth := setupDefaultAuth(client, evmAccount.key)
 
 		// Deploy
 		contractAddress, tx, counter, err := mock.DeployCounter(auth, client)
 		s.Require().NoError(err)
 
-		s.waitForTransaction(client, tx)
+		s.waitForTransaction(client, tx, evmAccount.address)
 		s.T().Logf("ContractAddress : %s", contractAddress.String())
 
 		// 6. Interact w/ contract and see changes
 		tx, err = counter.Increment(auth)
 		s.Require().NoError(err)
-		s.waitForTransaction(client, tx)
+		s.waitForTransaction(client, tx, evmAccount.address)
 
 		counterValue, err := counter.GetCounter(nil)
 		s.Require().NoError(err)
@@ -119,103 +100,21 @@ func (s *IntegrationTestSuite) testEVM(jsonRCP string) {
 	})
 }
 
-// setupEVMwithFunds sets up a new EVM account and sends funds from Alice to it, checking balance changes
-func (s *IntegrationTestSuite) setupEVMwithFunds(jsonRCP, chainEndpoint string, valIdx int) (*ecdsa.PrivateKey, common.Address) {
-	// 1. Create new account
-	// Make a key
-	key, err := crypto.HexToECDSA("88cbead91aee890d27bf06e003ade3d4e952427e88f88d31d61d3ef5e5d54305")
-	s.Require().NoError(err)
-
-	// Make a message to extract key, making sure we are using correct way
-	msg := crypto.Keccak256([]byte("foo"))
-	ethSig, _ := crypto.Sign(msg, key)
-	recoveredPub, _ := crypto.Ecrecover(msg, ethSig)
-
-	// Get pubkey, evm and cosmos address
-	pubKey, _ := crypto.UnmarshalPubkey(recoveredPub)
-	evmAddress := crypto.PubkeyToAddress(*pubKey)
-	s.T().Logf("Newly created evm address: %s", evmAddress)
-	cosmosAddress, err := PubKeyBytesToCosmosAddress(evmAddress.Bytes())
-	s.Require().NoError(err)
-	s.T().Logf("Newly created cosmos address: %s", cosmosAddress)
-
-	// Get alice's cosmos and evm address
-	alice, err := s.chainA.genesisAccounts[1].keyInfo.GetAddress()
-	s.Require().NoError(err)
-	s.T().Logf("Alice address: %s", alice)
-
-	publicKey, err := s.chainA.genesisAccounts[1].keyInfo.GetPubKey()
-	s.Require().NoError(err)
-	// Make sure we are using correct generation
-	aliceCosmosAddress, err := PubKeyToCosmosAddress(publicKey)
-	s.Require().NoError(err)
-	s.Require().Equal(alice.String(), aliceCosmosAddress)
-
-	// Get her EVM address
-	aliceEvmAddress, err := CosmosPubKeyToEVMAddress(publicKey)
-	s.Require().NoError(err)
-	s.T().Logf("Alice evm address : %s", aliceEvmAddress)
-
-	// 2. Send funds via cosmos for new account so it can do operations
-	s.execBankSend(s.chainA, valIdx, alice.String(), cosmosAddress, tokenAmount.String(), standardFees.String(), false)
-
-	var newBalance sdk.Coin
-	// Get balances of sender and recipient accounts
-	s.Require().Eventually(
-		func() bool {
-			// Get balance via cosmos
-			newBalance, err = getSpecificBalance(chainEndpoint, cosmosAddress, akiiDenom)
-			s.Require().NoError(err)
-
-			// Balance should already have some coin
-			return newBalance.IsValid() && newBalance.Amount.GT(math.ZeroInt())
-		},
-		10*time.Second,
-		5*time.Second,
-	)
-
-	s.Run("eth_getBalance on new address", func() {
-		// Get balance via evm
-		res, err := httpEVMPostJSON(jsonRCP, "eth_getBalance", []interface{}{
-			evmAddress.String(), "latest",
-		})
-		s.Require().NoError(err)
-
-		balance, err := parseResultAsHex(res)
-		s.Require().NoError(err)
-		s.T().Logf("Balance : %s", balance)
-
-		// Balance should have something
-		s.Require().False(strings.HasPrefix(balance, "0x0"))
-	})
-
-	// 3. Send via evm
-	client, err := ethclient.Dial(jsonRCP)
-	amount := big.NewInt(1000000000000000000)
-	receipt, err := sendEVM(client, key, evmAddress, aliceEvmAddress, amount)
-	s.Require().NoError(err)
-	s.T().Logf("Transaction status: %d\n", receipt.Status)
-
-	// 4. check changes
-	s.Run("eth_getBalance on address after send", func() {
-		res, err := httpEVMPostJSON(jsonRCP, "eth_getBalance", []interface{}{
-			evmAddress.String(), "latest",
-		})
-		s.Require().NoError(err)
-
-		balance, err := parseResultAsHex(res)
-		s.Require().NoError(err)
-		s.T().Logf("Balance : %s", balance)
-		// Balance should have something now
-		s.Require().False(strings.HasPrefix(balance, "0x0"))
-	})
-	return key, evmAddress
-}
-
-// waitForTransaction waits until transaction is mined, requiring its success
-func (s *IntegrationTestSuite) waitForTransaction(client *ethclient.Client, tx *geth.Transaction) *geth.Receipt {
+// waitForTransaction waits until transaction is mined, requiring its success and checks reason in case of failure
+func (s *IntegrationTestSuite) waitForTransaction(client *ethclient.Client, tx *geth.Transaction, sender common.Address) *geth.Receipt {
+	// Wait and check tx
 	receipt, err := bind.WaitMined(context.Background(), client, tx)
 	s.Require().NoError(err)
+
+	if receipt.Status == geth.ReceiptStatusFailed {
+		// Try to get the revert reason
+		reason, err := getRevertReason(client, tx.Hash(), sender)
+		if err != nil {
+			s.T().Logf("Failed to get revert reason: %v", err)
+		} else if reason != "" {
+			s.T().Logf("Revert reason: %s", reason)
+		}
+	}
 	s.Require().False(receipt.Status == geth.ReceiptStatusFailed)
 	return receipt
 }